feat(dunkel): move common table template to new folder
